feat: Access phone contacts.
Enabled the application to access contacts stored on the phone. Each
time the application is started, it checks if permission had been
granted for it to access contacts if, yes it gets those contacts and
display. Otherwise, it displays a dialog  telling the user to enable
access to contacts.

This feature brings with it lots of changes to how the app was before
now. Two packages; permission_handler and contacts_service are used to
to seek for permission and  access to contacts respectively. This is a
tad different from when a list was being used.

Moving forward, this will also bring more structural changes to the app
as the way contacts are being added and favorited will also change.